ref(ui): Move integration repository components
Just moving these out of the components folder since they aren't reused anywhere and I don't think they will be.
